seperated both modals into seperate componants
