+++ frontend/src/routes/(authenticated)/(dashboards)/admin/+page.svelte
@@ -5,25 +5,26 @@
   import Input from '$lib/forms/Input.svelte';
   import t from '$lib/i18n';
   import type { PageData } from './$types';
-  import { TrashIcon } from '$lib/icons';
   import IconButton from '$lib/components/IconButton.svelte';
-  import ButtonToggle from '$lib/components/ButtonToggle.svelte';
-  import { z } from 'zod';
-  import { FormModal } from '$lib/components/modals';
-  import { _changeUserAccountByAdmin } from './+page';
-  import type { ChangeUserAccountByAdminInput, LoadAdminDashboardQuery } from '$lib/gql/types';
   import DeleteUserModal from './DeleteUserModal.svelte';
-  import { hash } from '$lib/user';
+  import EditUserAccount from './EditUserAccount.svelte';
+  import type { LoadAdminDashboardQuery } from '$lib/gql/types';
+
+  type UserRow = LoadAdminDashboardQuery['users'][0]
 
   export let data: PageData;
   let deleteModal: DeleteUserModal;
+  let formModal: EditUserAccount;
+
   async function deleteUser(id: any): Promise<void>{
     formModal.close();
     await deleteModal.open(id);
   }
+  async function openModal(user: UserRow): Promise<void> {
+    await formModal.openModal(user);
+  }
   let projectSearch = '';
   let userSearch = '';
-  type UserRow = LoadAdminDashboardQuery['users'][0]
   $: projectSearchLower = projectSearch.toLocaleLowerCase();
   $: projects = data.projects
     .filter(
@@ -43,44 +44,6 @@
     )
     .slice(0, userSearch ? undefined : 10);
 
-  const schema = z.object({
-    email: z.string().email(),
-    name: z.string(),
-    password: z.string().optional(),
-    userId: z.string().optional(),
-  });
-  const verify = z.object({
-    keyphrase: z.string().refine((value) => value.match(`^${$t('admin_dashboard.enter_to_delete.user.value')}$`)),
-  });
-
-  let formModal: FormModal<typeof schema>;
-  $: form = formModal?.form();
-
-  async function openModal(user: UserRow): Promise<void> {
-    $form.email = user.email;
-    $form.name = user.name;
-    $form.userId = user.id;
-    await formModal.open(async () => {
-      if (data.user) {
-        const changeInput: ChangeUserAccountByAdminInput = {
-          adminId: data.user.id,
-          userId: user.id,
-          email: $form.email,
-          name: $form.name,
-        };
-        await _changeUserAccountByAdmin(changeInput);
-      }
-      let password: string = $form.password ?? '';
-      if (password !== '' && $form.password) {
-        await fetch('/api/Admin/resetPasswordAdmin', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ passwordHash: await hash(password), userId: user.id }),
-        });
-      }
-      return;
-    });
-  }
 </script>
 
 <svelte:head>
@@ -183,44 +146,6 @@
     </div>
   </div>
 
-  <FormModal bind:this={formModal} {schema} let:errors>
-    <span slot="title">{$t('admin_dashboard.form_modal.title')}</span>
-    <Input
-      id="email"
-      type="email"
-      label={$t('admin_dashboard.form_modal.email_label')}
-      bind:value={$form.email}
-      required
-      error={errors.email}
-      autofocus
-    />
-    <Input
-      id="name"
-      type="text"
-      label={$t('admin_dashboard.form_modal.name_label')}
-      bind:value={$form.name}
-      required
-      error={errors.name}
-      autofocus
-    />
-    <div class="text-error">
-      <Input
-        id="password"
-        type="password"
-        label={$t('admin_dashboard.form_modal.password_label')}
-        bind:value={$form.password}
-      />
-    </div>
-    <div style="display: flex" slot="extraActions" class="space-x-4">
-      <ButtonToggle theme="error" text1="unlock" text2="lock" icon1="i-mdi-lock" icon2="i-mdi-unlocked" />
-      <button
-        class="btn btn-error rounded"
-        on:click={async () => {
-          await deleteUser($form.userId);
-        }}>{$t('admin_dashboard.form_modal.delete_user')}<TrashIcon /></button
-      >
-    </div>
-    <span slot="submitText">{$t('admin_dashboard.form_modal.update_user')}</span>
-  </FormModal>
+  <EditUserAccount bind:this={formModal} deleteUser={deleteUser}></EditUserAccount>
   <DeleteUserModal bind:this={deleteModal}></DeleteUserModal>
 </main>
